requirements.txt
aiogram==3.4.1
python-dotenv==1.0.0
openai>=1.14.0
google-generativeai>=0.4.0
replicate>=0.30.0
requests>=2.31.0
aiohttp>=3.8.0
moviepy==1.0.3
Pillow>=10.0.0
imageio-ffmpeg>=0.4.5